Care_Facilities

This object a few types of health and community care facilities/services (Hospitals, Residential Care Facilities, Long-Term Care Homes, Community or Health Support Services, Independent Health Facilities, Pharmacies, and Primary Health Services) represented as spatial points.

data(Care_Facilities)

Format

A simple feature class (sf) point object containing 668 rows and 11 variables; each row represents a health/community care facility/service and thus addresses may be repeated if the location hosts multiple facilities/services.

Name

Name of facilities as specifed by the City of Hamilton or by Ontario Ministry of Health .

Service_Type

The type of service offered as described by the the Ontario Ministry of Health service provider locations feature layer documentation.

Service_Desc

The service as described by the the Ontario Ministry of Health service provider locations feature layer documentation.

Address

The address as supplied by the City of Hamilton or the Ontario Ministry of Health service provider locations feature layer documentation.

Address_2

Additional address details (unit number, floor, etc.).

Postalcode

Postal code.

Address_Desc

The address of the service may be either the "Pratice Location", "Administrative Office", "Administration and Practice Location", or blank. This field is specified by the the Ontario Ministry of Health service provider locations feature layer documentation.

Community

The community in which the service address is located in; either "Hamilton", "Burlington", "Grimbsy", or "NA".

Type

Type of facility, either "Community or Health Support Service", "Primary Health Service", "Fee-Based Independent Health Facility", "Hospital", "Laboratory - Specimen Collection Centre", or "Long-Term Care Home", "Pharmacy", or "Residential Care Facility" .

Supply

The number of 'supply' offered by the facility; for "Residential Care Facility" type this is the capacity range available at the location as specified by the City of Hamilton Open Data layer; all other types are unassigned (for now).

geometry

The sfc point geometry (location) of each facility/service.

Source

"Licensed residential care facilities" from Hamilton Open Data accessed November 10 2021; "Ministry of Health service provider locations" from Ontario GeoHub accessed November 14 2021.