City_Owned_Property

This object contains all city owned properties represented as as spatial points in the City of Hamilton.

data(City_Owned_Property)

Format

A simple feature class (sf) point object containing 1902 rows and 9 variables.

LOCATION

Approximate street address of the city-owned property land asset.

PROP_AREA

Area of the land asset in units of Acres.

WARD

Ward number of the land asset; this value aligns with "Ham_WardBound" object.

COMMUNITY

The current City of Hamilton consists of same six municipalities which form the Regional Municipality of Hamilton-Wentworth. These muncipalites are represented as 1) "An" for Ancaster, 2) "Du" for Dundas, 3) "Fl" for Flamborough, 4) "Gl" for Glanbrook, 5) "Ha" for Hamilton, and 6) "St" for Stoney Creek.

LEGAL_DESC

Description of the associated legal property document; likely for interal reference at the City.

CATEGORY

The category of the land asset, either: Parks and/or Open Space; Land; Utility; Open Space; Civic; Proposed Park; Recreation; Cemetery; Parking Lot; Cultural; Accessway; Road; Proposed Road; Utility. The city assigned these categories to each point and some include multiple categories (i.e. "Cemetery/Civc"). Use 'unique(City_Own_Prop$CATEGORY)' to see all combinations of categories.

ASSET_NAME

Name of the property as assigned by the City. ~41% of rows do not have an assigned name (NA).

geometry

The sfc point geometry (location) of each asset.

CATEGORY_TYPE

The type of category; this is an author assigned field grouping categories into 7 types namely: "Developed Outdoor" which corresponds to parks, open space, cementery, with civic infrastructure; "Underdeveloped Outdoor" which corresponds to cementeries with no infrastructure, flat land with mowed grass, proposed park space; "Remove" which corresponds to categories which are not destinations such as road, utility, accessway (an egress or ingress); "Natural/Open Space" which corresponds to natural areas, hydro lines, areas around storm management ponds, post-agricultural land; "Civic Service" which corresponds to Civic categories such as firestation, townhall, police station, airport; "Recreation/Cultural" which corresponds to locations such as library, historical buildings open to the public, golf courses, community centers, recreation centres, environmental education centres, stadiums; and "Parking Lots" which correspond to any properties with parking lots. See 00-Data-Preparation.Rmd for the specific categories which each category type is associated with.

Source

"City-Owned Properties" from Hamilton Open Data accessed November 8 2021.