Schools_201516_201011

A simple feature object containing all public and public catholic schools (elementary and secondary) in Hamilton, Ontario, Canada for the 2015-2016 and 2010-2011 academic years. Between the two years 154 schools remained in the same location, 3 moved (6 locations between these 3 schools), 3 were new in 2015-2016 (did not exist in 2010-2011), and 14 were removed in 2015-2016 (only existed in 2010-2011). Between the two years, 177 public and public catholic school locations existed in Hamilton, Ontario, Canada.

data(Schools_201516_201011)

Format

A simple feature (sf) point object containing 177 rows and 19 variables :

SchoolID

A unique character value for each school. Schools which are identical between the two years have the same ID but carry the suffix ".1011" or ".1516" depending on which year they are associated with.

Name

Name of school

CID2011

An integer ID corresponding to which the 2010-2011 catchment the school is located within; an ID associated with either a public elementary, public secondary, catholic elementary, or catholic secondary catchment.

CID2016

An integer ID corresponding to which the 2015-2016 catchment the school is located within; an ID associated with either a public elementary, public secondary, catholic elementary, or catholic secondary catchment.

System

A character tag identifying which school system the school belongs to (either "Public" or "Catholic")

Level

A characer tag identifying which school level the school instructs (either "High" or "Elementary"). Elementary indicates instruction of grades JK through 8 (approx. ages 5 to 13) and Secondary indicates instruction of grades 9 through 12 (approx. ages 14 to 18)

Year

A character tag identifying what year the information contained within the row corresponds to the school (either "2011 and 2016", "2011" or "2016" where 2011 and 2016 is shorthand for 2010-2011 and 2015-2016 respectively).

Status

A character tag identifying who the information corresponding to the school changed in that Year. The tag is either "NoChange" or "Expanded" aligning with the Year tag of "2011 and 2016"; "Moved" aligning with the Year tag of "2016" or "2011"; "New" aligning with the Year tag of "2016"; "Removed" aligning with the Year tag of "2011".

ON_Enrol_2011

The number of students historically enrolled in the school for the 2011-2012 academic year as sourced from the Ontario Open Data "Ontario public schools enrolment" dataset. Note, 2010-2011 is not available from the portal, so the oldest year available, 2011-2012, is used.

ON_Enrol_2016

The number of students historically enrolled in the school for the 2015-2016 academic year as sourced from the Ontario Open Data "Ontario public schools enrolment" dataset.

footprint2011

The footprint of the school building in the year 2010 in units of m^2 - derived from spatial polygons produced by DMTI Spatial Inc. using GeoScholars (retrieved January 17, 2021)

footprint2016

The footprint of the school buidling in the year 2016 in units of m^2 - derived from spatial polygons archieved by OpenStreetMaps (retrieved January 17, 2021)

ELEM

Binary character tag identifying if the school is an Elementary school (Grades JK to 8) or not ("ELEM" or "0")

MID

Binary character tag identifying if the school is a Middle school (Grades 6 to 8) or not ("MID" or "0"). Note: a Middle school is a type of Elementary school so all Middle schools are Elementary Schools but not all Elementary schools are Middle schools.

HIGH

Binary character tag identifying if the school is a Secondary school (Grades 9 to 12) or not ("HIGH" or "0")

geometry

The sf point geometry

urban.dist

Distance away (in units of 'm') from the centroid of the Central Buisness District (the lat-long coordinate of "King St W and James St S: (43.256684, -79.869039)). This distance is used as a 'proxy' for age of construction as schools further from the CBD are likely newer.

OTGC2011

Estimated on-the-ground-capacity (OTGC) for the 2010-2011 academic year - units of students

OTGC2016

Estimated on-the-ground-capacity (OTGC) for the 2015-2016 academic year - units of students

Source

School name, catchment ID, system, level, year, status are sourced from spatial points provided directly by the Hamilton Wentworth District School Board (HWDSB) and the Hamilton Wentworth Catholic District School Board (HWCDSB) for the respective academic years (2010-2011 and 2015-2016).

Historical Enrollment: sourced from "Ontario public schools enrolment" Ontario Open Data. Accessed: November 7, 2021

Footprint area: sourced from spatial polygons from DMTI Spatial Inc. using GeoScholars (retrieved January 17, 2021) and OpenStreetMaps (retrieved January 17, 2021)

Examples

 data(Schools_201516_201011)
 OTGC2011 <- Schools_201516_201011$OTGC2011